+++ CHANGELOG.md
@@ -6,13 +6,14 @@ and the project versioning adheres to [Semantic Versioning](https://semver.org/s
 
 ## [1.9.0] - UNRELEASED
 ### Added
+- [#103](https://github.com/serpro69/kotlin-faker/issues/103) [core] Add support for `Collection` types in `RandomProvider#randomClassInstance`
 - [#96](https://github.com/serpro69/kotlin-faker/issues/96) [core] Add `randomSubset` and `randomSublist` to `RandomService`
 - [#92](https://github.com/serpro69/kotlin-faker/issues/92) [core] Add `randomString` function to `RandomService`
 - [#86](https://github.com/serpro69/kotlin-faker/issues/86) [core] Generate birth-date based on the age
 
 ### Changed
-- [#97](https://github.com/serpro69/kotlin-faker/issues/97) [core] Change `RandomService#nextString` to generate strings only within given locale
 - [#100](https://github.com/serpro69/kotlin-faker/issues/100) [core] Add deprecation warning for `RandomService#nextString` since it's going to be replaced with `RandomService#randomString`
+- [#97](https://github.com/serpro69/kotlin-faker/issues/97) [core] Change `RandomService#nextString` to generate strings only within given locale
 
 ### Fixed
 - [#104](https://github.com/serpro69/kotlin-faker/issues/104) [core] `RandomProvider#randomClassInstance` : 'No suitable constructor found' for primitive classes